Link authors to search page
Link authors to search page for the same query, to allow finding books
by the same author.

Note: This doesn't work well for multiple authors because they are
stored as a single text column in the database, rather than as one
column per author or a many to many relationship. However, multi-author
fiction books are rare, so this isn't a major problem.