Format dates according to locale
ISO dates are convenient for storage and sorting but when displaying
dates they should be formatted according to the locale. We have used the
server locale as it is the default and is likely to line up with the
locale of most users given that the server is in the UK and this is a UK
book club.

Closes #15